Don't truncate scaled range
Some ranges are supposed to be fractional, so should not be truncated.
Previous change went too far and caused test failures.

+++ custom_components/tuya_local/helpers/device_config.py
@@ -44,7 +44,7 @@ def _scale_range(r, s):
     "Scale range r by factor s"
     if s == 1:
         return r
-    return {"min": int(r["min"] / s), "max": int(r["max"] / s)}
+    return {"min": r["min"] / s, "max": r["max"] / s}
 
 
 _unsigned_fmts = {
